main/700-799/785D_test.go
package main

import (
	"github.com/EndlessCheng/codeforces-go/main/testutil"
	"github.com/stretchr/testify/assert"
	"testing"
)

// https://codeforces.com/problemset/problem/785/D
// https://codeforces.com/problemset/status/785/problem/D
func TestCF785D(t *testing.T) {
	// just copy from website
	rawText := `
inputCopy
)(()()
outputCopy
6
inputCopy
()()()
outputCopy
7
inputCopy
)))
outputCopy
0`
	testutil.AssertEqualCase(t, rawText, 0, CF785D)
}

func TestCheckCF785D(t *testing.T) {
	return
	assert := assert.New(t)
	_ = assert

	inputGenerator := func() (string, testutil.OutputChecker) {
		//return ``
		rg := testutil.NewRandGenerator()
		rg.StrInSet(1, 5, "()")
		return rg.String(), func(myOutput string) (_b bool) {
			return true
		}
	}

	testutil.CheckRunResultsInf(t, inputGenerator, CF785D)
}
